Added support for multiple/uncombined search engine flags.
Previously, when passing arguments to the search engine via the "flags"
key in a pattern.json file. the argument would be treated as one single,
long argument by `exec.Command`.

This is fine when passing arguments like `"-rnwE"` (as this is
technically a single argument from the perspective of the shell),
however if "flags" was instead defined as `"-r -n -w -E"`, it would be
passed as one big argument to the search engine, causing errors
(e.g. `"invalid option"` in `grep` and `ag`).

This patch ensures that multiple arguments defined in a pattern.json
file are treated as seperate arguments (e.g. `-r`, `-n`, `-w`, `-E`).